Move position class into seprate file
